# apps/cautai-server/src/routes/search.ts
/**
 * @fileoverview Search API Routes
 * @author Cautai Team
 * @version 1.0.0
 */

import { FastifyInstance, FastifyPluginAsync } from 'fastify';
import { Static, Type } from '@sinclair/typebox';
import type { SearchResult, ComposeRequest, ComposeResponse } from '../config.js';

// Search request schemas
const SearchRequest = Type.Object({
  query: Type.String({ minLength: 1, maxLength: 500 }),
  limit: Type.Optional(Type.Number({ minimum: 1, maximum: 50, default: 10 })),
  offset: Type.Optional(Type.Number({ minimum: 0, default: 0 })),
  filters: Type.Optional(Type.Object({
    domain: Type.Optional(Type.String()),
    dateRange: Type.Optional(Type.Object({
      start: Type.String({ format: 'date' }),
      end: Type.String({ format: 'date' }),
    })),
    contentType: Type.Optional(Type.Union([
      Type.Literal('article'),
      Type.Literal('video'),
      Type.Literal('pdf'),
      Type.Literal('all'),
    ])),
  })),
});

const SearchResponse = Type.Object({
  results: Type.Array(Type.Object({
    url: Type.String(),
    title: Type.String(),
    snippet: Type.String(),
    domain: Type.String(),
    publishedAt: Type.Optional(Type.String()),
    score: Type.Number(),
    citations: Type.Array(Type.String()),
  })),
  total: Type.Number(),
  query: Type.String(),
  processingTimeMs: Type.Number(),
});

type SearchRequest = Static<typeof SearchRequest>;
type SearchResponse = Static<typeof SearchResponse>;

const searchRoutes: FastifyPluginAsync = async (fastify: FastifyInstance) => {
  // Web search endpoint
  fastify.post<{ Body: SearchRequest; Reply: SearchResponse }>(
    '/search',
    {
      preHandler: [fastify.authenticate],
      schema: {
        body: SearchRequest,
        response: {
          200: SearchResponse,
        },
      },
    },
    async (request) => {
      const startTime = Date.now();
      const { query, limit = 10, offset = 0, filters } = request.body;
      
      // Mock search implementation - replace with real search engine
      const mockResults = Array.from({ length: Math.min(limit, 5) }, (_, i) => ({
        url: `https://example.com/result-${i + offset + 1}`,
        title: `Search Result ${i + offset + 1} for "${query}"`,
        snippet: `This is a mock search result snippet for the query "${query}". It contains relevant information about the topic.`,
        domain: 'example.com',
        publishedAt: new Date(Date.now() - Math.random() * 86400000 * 30).toISOString(),
        score: Math.random() * 0.5 + 0.5,
        citations: [`Citation ${i + 1}`, `Source ${i + 1}`],
      }));
      
      return {
        results: mockResults,
        total: 42, // Mock total
        query,
        processingTimeMs: Date.now() - startTime,
      };
    }
  );
  
  // Compose answer endpoint
  fastify.post<{ Body: ComposeRequest; Reply: ComposeResponse }>(
    '/compose',
    {
      preHandler: [fastify.authenticate],
      schema: {
        body: Type.Object({
          query: Type.String(),
          sources: Type.Array(Type.Object({
            url: Type.String(),
            title: Type.String(),
            snippet: Type.String(),
          })),
          style: Type.Optional(Type.Union([
            Type.Literal('concise'),
            Type.Literal('detailed'),
            Type.Literal('academic'),
          ])),
        }),
        response: {
          200: Type.Object({
            answer: Type.String(),
            sources: Type.Array(Type.String()),
            confidence: Type.Number(),
            processingTimeMs: Type.Number(),
          }),
        },
      },
    },
    async (request) => {
      const startTime = Date.now();
      const { query, sources, style = 'detailed' } = request.body;
      
      // Mock compose implementation
      const answer = `Based on the search results for "${query}", here is a comprehensive answer:\n\n` +
        `This is a mock composed answer that would synthesize information from the provided sources. ` +
        `The answer would be generated in ${style} style and provide accurate information based on ` +
        `the ${sources.length} sources provided.`;
      
      return {
        answer,
        sources: sources.map(s => s.url),
        confidence: 0.85,
        processingTimeMs: Date.now() - startTime,
      };
    }
  );
  
  // Get citations endpoint
  fastify.get<{
    Querystring: { url: string };
    Reply: { citations: Array<{ text: string; source: string; confidence: number }> };
  }>(
    '/citations',
    {
      preHandler: [fastify.authenticate],
      schema: {
        querystring: Type.Object({
          url: Type.String({ format: 'uri' }),
        }),
        response: {
          200: Type.Object({
            citations: Type.Array(Type.Object({
              text: Type.String(),
              source: Type.String(),
              confidence: Type.Number(),
            })),
          }),
        },
      },
    },
    async (request) => {
      const { url } = request.query;
      
      // Mock citations implementation
      return {
        citations: [
          {
            text: 'Mock citation text from the provided URL',
            source: url,
            confidence: 0.92,
          },
          {
            text: 'Another relevant citation extracted from the content',
            source: url,
            confidence: 0.87,
          },
        ],
      };
    }
  );
};

export { searchRoutes };