# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cautai AI Search Engine API
  description: |
    Cautai is an AI-first search engine designed for both agents and humans. This API provides 
    comprehensive search capabilities, content composition, and citation generation services.
    
    ## Features
    - Advanced search with hybrid ranking (BM25 + semantic similarity)
    - AI-powered content composition and summarization
    - Academic citation generation in multiple formats
    - Multi-language support (English, Romanian)
    - Rate limiting and authentication
    - Real-time search result processing
    
    ## Authentication
    Most endpoints require Bearer token authentication. Obtain tokens through the authentication endpoints.
    
    ## Rate Limiting
    - Search endpoints: 100 requests per minute per IP
    - Composition endpoints: 20 requests per minute per authenticated user
    - Citation endpoints: 50 requests per minute per authenticated user
    
  version: 1.0.0
  contact:
    name: Cautai Support
    url: https://cautai.ro/support
    email: support@cautai.ro
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.cautai.ro/v1
    description: Production server
  - url: https://staging-api.cautai.ro/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
  /search:
    post:
      summary: Perform search query
      description: |
        Execute a search query against multiple sources with hybrid ranking and filtering options.
        Results are ranked using a combination of BM25 and semantic similarity algorithms.
      operationId: search
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              basic_search:
                summary: Basic search query
                value:
                  query: "machine learning algorithms"
                  limit: 10
                  language: "en"
              advanced_search:
                summary: Advanced search with filters
                value:
                  query: "quantum computing applications"
                  limit: 20
                  language: "en"
                  contentType: "academic"
                  dateRange: "year"
                  safeSearch: true
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '408':
          description: Search timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
          description: Search query string
          example: "machine learning algorithms"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Maximum number of results to return
        offset:
          type: integer
          minimum: 0
          default: 0
          description: Number of results to skip (for pagination)
        language:
          type: string
          enum: ["en", "ro"]
          default: "en"
          description: Language preference for results
        contentType:
          type: string
          enum: ["web", "academic", "news", "blog"]
          description: Filter by content type
        dateRange:
          type: string
          enum: ["day", "week", "month", "year", "all"]
          description: Time range for content filtering
        safeSearch:
          type: boolean
          default: true
          description: Enable safe search filtering
        timeout:
          type: integer
          minimum: 1000
          maximum: 30000
          default: 10000
          description: Search timeout in milliseconds
    SearchResponse:
      type: object
      properties:
        searchId:
          type: string
          format: uuid
          description: Unique identifier for this search
        query:
          type: string
          description: Original search query
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
          description: Total number of results found
        processingTime:
          type: integer
          description: Processing time in milliseconds
        hasMore:
          type: boolean
          description: Whether more results are available
        metadata:
          type: object
          properties:
            filters:
              type: object
              description: Applied filters
            sources:
              type: array
              items:
                type: string
              description: Search sources used
            ranking:
              type: object
              description: Ranking algorithm information
    SearchResult:
      type: object
      properties:
        id:
          type: string
          description: Unique result identifier
        title:
          type: string
          description: Result title
        url:
          type: string
          format: uri
          description: Source URL
        snippet:
          type: string
          description: Content excerpt
        relevance:
          type: number
          minimum: 0
          maximum: 1
          description: Relevance score
        source:
          type: string
          description: Source identifier
        timestamp:
          type: string
          format: date-time
          description: When the result was indexed
        metadata:
          type: object
          description: Additional result metadata
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          description: Unique request identifier for debugging
...